rework how compiling text files works
