+++ game/assets/jak1/subtitle/game_subtitle_es.gs
