+++ game/assets/jak1/text/game_text_en.gs
