+++ goal_src/pc/pckernel-h.gc
@@ -23,9 +23,6 @@
  |#
 
 
-(#when PC_PORT
-
-
 ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
 ;;;; constants
 ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
@@ -578,9 +575,6 @@
       )
     )
   ))
-
-
-)
