+++ goal_src/pc/pckernel.gc
@@ -22,9 +22,6 @@
 
  |#
 
-;; is this necessary?
-(#when PC_PORT
-
 
 ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
 ;;;; updates
@@ -802,9 +799,7 @@
                           0
                           (* 10 (-> *video-parms* relative-y-scale))
                           (font-color flat-yellow)
-                          (font-flags kerning))))
-
-) ;; when PC_PORT
+                          (font-flags shadow kerning))))
