+++ goal_src/pc/subtitle.gc
@@ -16,15 +16,12 @@
  |#
 
 
-(#when PC_PORT
-
-
 ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
 ;;;; constants
 ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
 
 
-(defconstant PC_SUBTITLE_FILE_SIZE (* 64 1024))
+(defconstant PC_SUBTITLE_FILE_SIZE (* 64 1024)) ;; 64K heap for subtitles. adjust later if necessary.
 (defconstant PC_SUBTITLE_FILE_NAME "subtit")
 (defglobalconstant PC_SUBTITLE_DEBUG #f)
 
@@ -250,7 +247,7 @@
 (defmacro reload-subtitles ()
   "rebuild and reload subtitles."
   `(begin
-      (asm-text-file subtitle (game-text-version jak1-v1) :files ("game/assets/game_subtitle.gp"))
+      (asm-text-file subtitle :files ("game/assets/game_subtitle.gp"))
       (+! (-> *subtitle-text* lang) (the-as pc-subtitle-lang 1))))
 
 
@@ -722,9 +719,6 @@
 
 ;; start the subtitle process
 (subtitle-start)
-
-
-)
