+++ goalc/CMakeLists.txt
@@ -27,11 +27,10 @@ add_library(compiler
         compiler/compilation/State.cpp
         compiler/compilation/Static.cpp
         compiler/Util.cpp
-        data_compiler/game_text.cpp
+        data_compiler/game_text_common.cpp
         data_compiler/dir_tpages.cpp
         data_compiler/DataObjectGenerator.cpp
         data_compiler/game_count.cpp
-        data_compiler/game_subtitle.cpp
         debugger/Debugger.cpp
         debugger/DebugInfo.cpp
         listener/Listener.cpp
