+++ test/test_data/test_game_text.gp
@@ -1,0 +1,2 @@
+(text
+  (jak1-v1 "test/test_data/test_game_text.txt"))
